// frontend/web/src/services/API/proserve-wb-projects-api/models/GetProjectAccountsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Projects API
 * API in charge of processing the methods required for the use cases
 *
 * The version of the OpenAPI document: 2024-04-11T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ProjectAccount } from './ProjectAccount';
import {
    ProjectAccountFromJSON,
    ProjectAccountFromJSONTyped,
    ProjectAccountToJSON,
} from './ProjectAccount';

/**
 * 
 * @export
 * @interface GetProjectAccountsResponse
 */
export interface GetProjectAccountsResponse {
    /**
     * 
     * @type {Array<ProjectAccount>}
     * @memberof GetProjectAccountsResponse
     */
    projectAccounts: Array<ProjectAccount>;
    /**
     * 
     * @type {object}
     * @memberof GetProjectAccountsResponse
     */
    nextToken?: object | null;
}

/**
 * Check if a given object implements the GetProjectAccountsResponse interface.
 */
export function instanceOfGetProjectAccountsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "projectAccounts" in value;

    return isInstance;
}

export function GetProjectAccountsResponseFromJSON(json: any): GetProjectAccountsResponse {
    return GetProjectAccountsResponseFromJSONTyped(json, false);
}

export function GetProjectAccountsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProjectAccountsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'projectAccounts': ((json['projectAccounts'] as Array<any>).map(ProjectAccountFromJSON)),
        'nextToken': !exists(json, 'nextToken') ? undefined : json['nextToken'],
    };
}

export function GetProjectAccountsResponseToJSON(value?: GetProjectAccountsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'projectAccounts': ((value.projectAccounts as Array<any>).map(ProjectAccountToJSON)),
        'nextToken': value.nextToken,
    };
}