// frontend/web/src/services/API/proserve-wb-projects-api/models/GetProjectsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Projects API
 * API in charge of processing the methods required for the use cases
 *
 * The version of the OpenAPI document: 2024-04-11T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Project } from './Project';
import {
    ProjectFromJSON,
    ProjectFromJSONTyped,
    ProjectToJSON,
} from './Project';
import type { ProjectAssignment } from './ProjectAssignment';
import {
    ProjectAssignmentFromJSON,
    ProjectAssignmentFromJSONTyped,
    ProjectAssignmentToJSON,
} from './ProjectAssignment';
import type { ProjectEnrolment } from './ProjectEnrolment';
import {
    ProjectEnrolmentFromJSON,
    ProjectEnrolmentFromJSONTyped,
    ProjectEnrolmentToJSON,
} from './ProjectEnrolment';

/**
 * 
 * @export
 * @interface GetProjectsResponse
 */
export interface GetProjectsResponse {
    /**
     * 
     * @type {object}
     * @memberof GetProjectsResponse
     */
    nextToken?: object | null;
    /**
     * 
     * @type {Array<Project>}
     * @memberof GetProjectsResponse
     */
    projects: Array<Project>;
    /**
     * 
     * @type {Array<ProjectAssignment>}
     * @memberof GetProjectsResponse
     */
    assignments?: Array<ProjectAssignment>;
    /**
     * 
     * @type {Array<ProjectEnrolment>}
     * @memberof GetProjectsResponse
     */
    enrolments?: Array<ProjectEnrolment>;
}

/**
 * Check if a given object implements the GetProjectsResponse interface.
 */
export function instanceOfGetProjectsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "projects" in value;

    return isInstance;
}

export function GetProjectsResponseFromJSON(json: any): GetProjectsResponse {
    return GetProjectsResponseFromJSONTyped(json, false);
}

export function GetProjectsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProjectsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'nextToken': !exists(json, 'nextToken') ? undefined : json['nextToken'],
        'projects': ((json['projects'] as Array<any>).map(ProjectFromJSON)),
        'assignments': !exists(json, 'assignments') ? undefined : ((json['assignments'] as Array<any>).map(ProjectAssignmentFromJSON)),
        'enrolments': !exists(json, 'enrolments') ? undefined : ((json['enrolments'] as Array<any>).map(ProjectEnrolmentFromJSON)),
    };
}

export function GetProjectsResponseToJSON(value?: GetProjectsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'nextToken': value.nextToken,
        'projects': ((value.projects as Array<any>).map(ProjectToJSON)),
        'assignments': value.assignments === undefined ? undefined : ((value.assignments as Array<any>).map(ProjectAssignmentToJSON)),
        'enrolments': value.enrolments === undefined ? undefined : ((value.enrolments as Array<any>).map(ProjectEnrolmentToJSON)),
    };
}