// frontend/web/src/services/API/proserve-wb-projects-api/models/OnBoardProjectAccountRequest.ts
export interface OnBoardProjectAccountRequest {
    /**
     * Unique ID of the account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    id?: string;
    /**
     * Unique ID of the AWS account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    awsAccountId: string;
    /**
     * Type of the account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    accountType: string;
    /**
     * Name of the account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    accountName?: string;
    /**
     * Description of the account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    accountDescription?: string;
    /**
     * Point in time where account is created.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    createDate?: string;
    /**
     * Point in time where account is last updated.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    lastUpdateDate?: string;
    /**
     * Account status
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    accountStatus?: string;
    /**
     * The stage account represents. Possible values dev, qa, prod
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    stage?: string;
    /**
     * The reference to the associated technology platform
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    technologyId: string;
    /**
     * Region in which workbenches will be provisioned in this account.
     * @type {string}
     * @memberof OnBoardProjectAccountRequest
     */
    region: string;
}