// frontend/web/src/services/API/proserve-wb-projects-api/models/GetProjectAssignmentsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Projects API
 * API in charge of processing the methods required for the use cases
 *
 * The version of the OpenAPI document: 2024-04-11T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { GetProjectAssignmentsResponseItem } from './GetProjectAssignmentsResponseItem';
import {
    GetProjectAssignmentsResponseItemFromJSON,
    GetProjectAssignmentsResponseItemFromJSONTyped,
    GetProjectAssignmentsResponseItemToJSON,
} from './GetProjectAssignmentsResponseItem';

/**
 * 
 * @export
 * @interface GetProjectAssignmentsResponse
 */
export interface GetProjectAssignmentsResponse {
    /**
     * 
     * @type {string}
     * @memberof GetProjectAssignmentsResponse
     */
    nextToken?: string | null;
    /**
     * 
     * @type {Array<GetProjectAssignmentsResponseItem>}
     * @memberof GetProjectAssignmentsResponse
     */
    assignments?: Array<GetProjectAssignmentsResponseItem>;
}

/**
 * Check if a given object implements the GetProjectAssignmentsResponse interface.
 */
export function instanceOfGetProjectAssignmentsResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetProjectAssignmentsResponseFromJSON(json: any): GetProjectAssignmentsResponse {
    return GetProjectAssignmentsResponseFromJSONTyped(json, false);
}

export function GetProjectAssignmentsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProjectAssignmentsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'nextToken': !exists(json, 'nextToken') ? undefined : json['nextToken'],
        'assignments': !exists(json, 'assignments') ? undefined : ((json['assignments'] as Array<any>).map(GetProjectAssignmentsResponseItemFromJSON)),
    };
}

export function GetProjectAssignmentsResponseToJSON(value?: GetProjectAssignmentsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'nextToken': value.nextToken,
        'assignments': value.assignments === undefined ? undefined : ((value.assignments as Array<any>).map(GetProjectAssignmentsResponseItemToJSON)),
    };
}