// frontend/web/src/services/API/proserve-wb-projects-api/models/AssignUserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Projects API
 * API in charge of processing the methods required for the use cases
 *
 * The version of the OpenAPI document: 2024-04-11T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AssignUserRequest
 */
export interface AssignUserRequest {
    /**
     * User TID
     * @type {string}
     * @memberof AssignUserRequest
     */
    userId?: string;
    /**
     * User roles in the project
     * @type {Array<string>}
     * @memberof AssignUserRequest
     */
    roles?: Array<string>;
}

/**
 * Check if a given object implements the AssignUserRequest interface.
 */
export function instanceOfAssignUserRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function AssignUserRequestFromJSON(json: any): AssignUserRequest {
    return AssignUserRequestFromJSONTyped(json, false);
}

export function AssignUserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AssignUserRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'roles': !exists(json, 'roles') ? undefined : json['roles'],
    };
}

export function AssignUserRequestToJSON(value?: AssignUserRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userId': value.userId,
        'roles': value.roles,
    };
}